refactor: 구조 변경
step4 진행을 위하여 기존 step3 패키지 이름 및 패키지 구조 변경
