+++ src/test/java/study/racingcar/RacingCarGameTest.java
@@ -1,11 +1,12 @@
-package study.step3;
+package study.racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import study.racingcar.entity.RacingCar;
+import study.racingcar.util.RandomGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
